refactor: exception 레이어 이동
