+++ company/src/main/java/com/sparta/hotdeal/company/application/service/CompanyService.java
@@ -7,8 +7,8 @@ import com.sparta.hotdeal.company.application.dtos.company.ResGetCompanyByIdDto;
 import com.sparta.hotdeal.company.domain.entity.company.Company;
 import com.sparta.hotdeal.company.domain.entity.company.CompanyStatusEnum;
 import com.sparta.hotdeal.company.domain.repository.CompanyRepository;
-import com.sparta.hotdeal.company.infrastructure.exception.ApplicationException;
-import com.sparta.hotdeal.company.infrastructure.exception.ErrorCode;
+import com.sparta.hotdeal.company.application.exception.ApplicationException;
+import com.sparta.hotdeal.company.application.exception.ErrorCode;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
